requirements.txt
selenium==4.31.0   # released 2025‑04‑05
numpy==2.2.4       # released 2025‑03‑16
Pillow==11.2.1     # released 2025‑04‑12
python-dotenv>=1.1
